+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Fixed
 - Improved handling UFG labels for iOS
+- Improved handling errors from Android UFG lib
 
 ## [3.213.3-beta] - 2022-03-28
 ### Fixed